+++ app/src/frontend/building/data-containers/construction.tsx
@@ -10,6 +10,8 @@ import { CategoryViewProps } from './category-view-props';
 import InfoBox from '../../components/info-box';
 import { DataEntryGroup } from '../data-components/data-entry-group';
 
+import {t} from'i18next';
+
 const ConstructionMaterialsOptions = [
     'Wood',
     'Stone',
@@ -38,12 +40,12 @@ const RoofCoveringOptions = [
 const ConstructionView: React.FunctionComponent<CategoryViewProps> = (props) => {
     return (
          <Fragment>
-            <DataEntryGroup name="Materials">
+            <DataEntryGroup name={t("Materials")}>
                 <SelectDataEntry
-                    title={dataFields.construction_core_material.title}
+                    title={t(dataFields.construction_core_material.title)}
                     slug="construction_core_material"
                     value={props.building.construction_core_material}
-                    tooltip={dataFields.construction_core_material.tooltip}
+                    tooltip={t(dataFields.construction_core_material.tooltip)}
                     options={ConstructionMaterialsOptions}
                     mode={props.mode}
                     copy={props.copy}
@@ -58,21 +60,21 @@ const ConstructionView: React.FunctionComponent<CategoryViewProps> = (props) => 
                     verified_count={props.building.verified.construction_core_material}
                     />
                 <SelectDataEntry
-                    title={dataFields.construction_secondary_materials.title}
+                    title={t(dataFields.construction_secondary_materials.title)}
                     disabled={true}
                     slug="construction_secondary_materials"
                     value={props.building.construction_secondary_materials}
-                    tooltip={dataFields.construction_secondary_materials.tooltip}
+                    tooltip={t(dataFields.construction_secondary_materials.tooltip)}
                     options={ConstructionMaterialsOptions}
                     mode={props.mode}
                     copy={props.copy}
                     onChange={props.onChange}
                 />
                 <SelectDataEntry
-                    title={dataFields.construction_roof_covering.title}
+                    title={t(dataFields.construction_roof_covering.title)}
                     slug="construction_roof_covering"
                     value={props.building.construction_roof_covering}
-                    tooltip={dataFields.construction_roof_covering.tooltip}
+                    tooltip={t(dataFields.construction_roof_covering.tooltip)}
                     options={RoofCoveringOptions}
                     mode={props.mode}
                     copy={props.copy}
@@ -89,7 +91,7 @@ const ConstructionView: React.FunctionComponent<CategoryViewProps> = (props) => 
             </DataEntryGroup>
             <DataEntryGroup name="Construction sectors">
                 <DataEntry
-                    title="Construction system type"
+                    title={t("Construction system type")}
                     slug=""
                     value=""
                     mode='view'