+++ app/src/frontend/building/header-buttons/view-edit-control.tsx
@@ -4,6 +4,8 @@ import { NavLink } from 'react-router-dom';
 import { EditIcon, ViewIcon } from '../../components/icons';
 import { Building } from '../../models/building';
 
+import {t} from'i18next';
+
 interface ViewEditControlProps {
     cat: string;
     mode: 'view' | 'edit';
@@ -14,16 +16,16 @@ const ViewEditControl: React.FC<ViewEditControlProps> = props => (
     (props.mode === 'edit')?
                         <NavLink
                             className="icon-button view"
-                            title="View data"
+                            title={t("View data")}
                             to={`/view/${props.cat}/${props.building.building_id}`}>
-                            View
+                            {t("View")}
                             <ViewIcon />
                         </NavLink>
                         : <NavLink
                             className="icon-button edit"
-                            title="Edit data"
+                            title={t("Edit data")}
                             to={`/edit/${props.cat}/${props.building.building_id}`}>
-                            Edit
+                            {t("Edit")}
                             <EditIcon />
                         </NavLink>
 );